Refactor PlayerSearchController and move methods to it
